refactor: move the react ui components into reusable folders
