+++ apps/react-ui/client/src/components/ThemeToggle.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { useTheme } from "../providers"
+import { useTheme } from "../providers/ThemeProvider"
 
 export function ThemeToggle() {
 	const { theme, toggleTheme } = useTheme()
