+++ apps/react-ui/client/src/providers/index.tsx
@@ -1,0 +1,7 @@
+// src/providers/index.tsx
+import React from "react"
+import { ThemeProvider } from "./ThemeProvider"
+
+export const Providers = ({ children }: { children: React.ReactNode }) => {
+	return <ThemeProvider>{children}</ThemeProvider>
+}
