+++ apps/react-ui/client/src/styles/favicon.ico
